Major configHandler update to centralize config keys as attributes
Previously, each class pulled values from configHandler, making it
difficult to check for consistency and typing before the program ran.

This commit makes it so the config dict is loaded into the ConfigHandler
class' attributes, which are then assigned accessed by other classes as
necessary.

The get() method was also updated to include typechecking and value
checking.

The config class now also uses the Logger class.