remove PackedBed.entities -> Bead.tag
packedBed entities/tags/dimtags are now directly accessed from the
beads list.

Refactoring responsibility for the beads back to the bead class is a
good idea. In the future, I'd like to push generation to the Bead class
as well since we have to individually generate the beads anyway. That
way, the tag can be updated automatically during generation/copying.