Update copy_mesh to try with parallelism
NOTE: Parallelism using vanilla gmsh in the way I've written it is
impossible. Every `addNodes()` destroys the mesh cache and forces
rebuilding of the mesh cache upon `getVertexTag()` which is in
`addElements()`. Both of these operations are `critical` or `atomic`.

We would need to modify code within GMSH to get this to work.