Update copy mesh code to work with a modified GMSH
The issue with the old copymesh is that `addNodes()` and `addElements()`
had serial (atomic/critical) parts to it. This meant that copying meshes
in a loop would be insanely slow.

At the end of `addNodes()`, the mesh cache was destroyed. This is a
serial operation. Within `addElements()` there's a `getVertex..()` or
something that rebuilds the mesh cache when needed. Again, serial.

Now we can use a modified GMSH with custom addNodes() and addElements()
and instead of fully copying all nodes and elements of a mesh at once
for each copy, we copy all nodes, destroy cache, then copy all elements.

This makes it extremely fast. Roughly, what took 8-12 hours now only
takes 15 or so minutes.